+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 * Fixed an issue with the **format** command when contributing via the UI
 * The **format** command will now not remove the `defaultRows` key from incident, indicator and generic fields with `type: grid`.
 * Fixed an issue with the **validate** command when a layoutscontainer did not have the `fromversion` field set.
+* added a new command **update-xsoar-config-file** to handle your XSOAR Configuration File.
 * Added `skipVerify` argument in **upload** command to skip pack signature verification.
+* Adding a filter to the current user to `_get_playground_id` method to find the actual playground id.
 
 # 1.5.3
 * The **format** command will now set `unsearchable: True` for incident, indicator and generic fields.